core/requirements/requirements.txt
loguru
python-dotenv
httpx
sudo